+++ src/edu/gatech/lbs/sim/tracegenerator/paramdistribution/GaussianParamDistribution.java
@@ -1,4 +1,4 @@
-// Copyright (c) 2009, Georgia Tech Research Corporation
+// Copyright (c) 2012, Georgia Tech Research Corporation
 // Authors:
 //   Peter Pesti (pesti@gatech.edu)
 //
@@ -11,14 +11,14 @@ import edu.gatech.lbs.core.vector.IVector;
 public class GaussianParamDistribution implements IParamDistribution {
   public static final String xmlName = "gaussian";
 
-  private double mean; // [m] or [m/s] or [s]
-  private double stdev; // [m] or [m/s] or [s]
-  private double min; // [m] or [m/s] or [s]
-  private double max; // [m] or [m/s] or [s]
+  private int mean; // [mm] or [mm/s] or [ms]
+  private int stdev; // [mm] or [mm/s] or [ms]
+  private int min; // [mm] or [mm/s] or [ms]
+  private int max; // [mm] or [mm/s] or [ms]
 
   private Random rnd;
 
-  public GaussianParamDistribution(double mean, double stdev, double min, double max) {
+  public GaussianParamDistribution(int mean, int stdev, int min, int max) {
     this.mean = mean;
     this.stdev = stdev;
     this.min = min;
@@ -27,10 +27,10 @@ public class GaussianParamDistribution implements IParamDistribution {
     rnd = new Random();
   }
 
-  public double getNextValue(IVector location) {
-    double value;
+  public int getNextValue(IVector location) {
+    int value;
     do {
-      value = mean + rnd.nextGaussian() * stdev;
+      value = (int) (mean + rnd.nextGaussian() * stdev);
     } while (value <= min || value > max);
 
     return value;
